- warning level 4.
git-svn-id: http://micro-profiler.googlecode.com/svn/trunk@206 1676f2fc-95af-a39c-d699-6f41337c5309

+++ libraries/os/mt.h
@@ -13,7 +13,7 @@ namespace os
 	{
 		enum wait_status { satisfied, timeout, abandoned };
 
-		static const unsigned int infinite = -1;
+		static const unsigned int infinite = static_cast<unsigned int>(-1);
 
 		virtual wait_status wait(unsigned int timeout) volatile = 0;
 	};
